--- target-api-react/src/client/authUtils.ts
@@ -1,144 +1,0 @@
-/*
- * (c) Copyright 2024 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export function generateCodeVerifier(): string {
-  // 43-128 characters, URL-safe base64
-  const array = new Uint8Array(64);
-  window.crypto.getRandomValues(array);
-  return btoa(String.fromCharCode(...array))
-    .replace(/\+/g, '-').replace(/\//g, '_').replace(/=+$/, '');
-}
-
-export async function generateCodeChallenge(codeVerifier: string): Promise<string> {
-  const encoder = new TextEncoder();
-  const data = encoder.encode(codeVerifier);
-  const digest = await window.crypto.subtle.digest('SHA-256', data);
-  const base64Digest = btoa(String.fromCharCode(...new Uint8Array(digest)))
-    .replace(/\+/g, '-').replace(/\//g, '_').replace(/=+$/, '');
-  return base64Digest;
-}
-
-// Function to open browser and get authorization code
-export function openBrowserAndGetAuthCode(
-  authUrl: string,
-  redirectUrl: string
-): Promise<string> {
-  return new Promise((resolve, reject) => {
-    // Open the authorization URL in a popup window
-    const popup = window.open(
-      authUrl,
-      "oauth-popup",
-      "width=500,height=600,scrollbars=yes,resizable=yes"
-    );
-
-    if (!popup) {
-      reject(
-        new Error(
-          "Failed to open popup window. Please allow popups for this site."
-        )
-      );
-      return;
-    }
-
-    // Check if the popup is closed manually by the user
-    const checkClosed = setInterval(() => {
-      if (popup.closed) {
-        clearInterval(checkClosed);
-        reject(new Error("Authentication was cancelled by the user."));
-      }
-    }, 1000);
-
-    // Listen for messages from the popup (if using postMessage approach)
-    const messageListener = (event: MessageEvent) => {
-      // Verify the origin for security
-      if (event.origin !== window.location.origin) {
-        return;
-      }
-
-      if (event.data.type === "OAUTH_SUCCESS" && event.data.code) {
-        clearInterval(checkClosed);
-        window.removeEventListener("message", messageListener);
-        popup.close();
-        resolve(event.data.code);
-      } else if (event.data.type === "OAUTH_ERROR") {
-        clearInterval(checkClosed);
-        window.removeEventListener("message", messageListener);
-        popup.close();
-        reject(new Error(event.data.error || "Authentication failed"));
-      }
-    };
-
-    window.addEventListener("message", messageListener);
-
-    // Poll the popup URL to check for redirect (fallback method)
-    const pollTimer = setInterval(() => {
-      try {
-        if (popup.location.href.includes(redirectUrl)) {
-          const url = new URL(popup.location.href);
-          const code = url.searchParams.get("code");
-          const error = url.searchParams.get("error");
-
-          clearInterval(pollTimer);
-          clearInterval(checkClosed);
-          window.removeEventListener("message", messageListener);
-          popup.close();
-
-          if (error) {
-            reject(new Error(`OAuth error: ${error}`));
-          } else if (code) {
-            resolve(code);
-          } else {
-            reject(new Error("No authorization code received"));
-          }
-        }
-      } catch (e) {
-        // Cross-origin error is expected while the popup is on the auth domain
-        // We'll continue polling until it redirects back to our domain
-      }
-    }, 1000);
-
-    // Set a timeout for the authentication process
-    setTimeout(() => {
-      clearInterval(pollTimer);
-      clearInterval(checkClosed);
-      window.removeEventListener("message", messageListener);
-      if (!popup.closed) {
-        popup.close();
-      }
-      reject(new Error("Authentication timeout. Please try again."));
-    }, 300000); // 5 minute timeout
-  });
-}
-
-
-// Token storage and retrieval
-export function saveToken(token: string, expiresIn: number) {
-  const expiresAt = Date.now() + expiresIn * 1000 - 5000; // 5s buffer
-  localStorage.setItem('auth_token', token);
-  localStorage.setItem('auth_token_expiry', expiresAt.toString());
-}
-
-export function getToken(): string | null {
-  const token = localStorage.getItem('auth_token');
-  const expiry = localStorage.getItem('auth_token_expiry');
-  if (!token || !expiry) return null;
-  if (Date.now() > parseInt(expiry, 10)) {
-    localStorage.removeItem('auth_token');
-    localStorage.removeItem('auth_token_expiry');
-    return null;
-  }
-  return token;
-}
